Modifying functions, tests, documentation

+++ PlanApp/src/main/java/planapp/ui/PlanAppUi.java
@@ -6,6 +6,7 @@ import java.io.InputStreamReader;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.Properties;
 import javafx.application.Application;
 import javafx.event.ActionEvent;
@@ -13,7 +14,10 @@ import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.geometry.Rectangle2D;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.Hyperlink;
 import javafx.scene.control.Label;
@@ -39,7 +43,6 @@ import planapp.domain.PlanService;
  * Application user interface
  */
 public class PlanAppUi extends Application {
-    
     private PlanService planService;
     private Scene plan;
     private Scene login;
@@ -68,7 +71,7 @@ public class PlanAppUi extends Application {
         Text welcome = new Text();
         
         logTitle.setFont(Font.font("Tahoma", FontWeight.NORMAL, 20));
-        logTitle.setText("PlanApp");    
+        logTitle.setText("PlanApp (ETSI)");    
         
         Hyperlink registerLink = new Hyperlink("Register");
         TextField logUsername = new TextField();
@@ -207,18 +210,23 @@ public class PlanAppUi extends Application {
         
         
         // Plan components:       
-        List<Course> courses = planService.allCourses();
-        List<Course> nextCourses = new ArrayList<>();
-        
         Text planNotification = new Text("");
+        Text planNotification1 = new Text("");
+                
+        ScrollPane planFlex = new ScrollPane();
+        ScrollPane suggFlex = new ScrollPane();
         
-        BorderPane planGrid = new BorderPane();      
+        BorderPane planGrid = new BorderPane(planFlex);      
         BorderPane topBar = new BorderPane();
+        
+        GridPane buttonBar = new GridPane();
+        buttonBar.setPadding(new Insets(10, 10, 10, 10));
         
         AnchorPane actions = new AnchorPane();
         HBox links = new HBox();
         Hyperlink logout = new Hyperlink("Logout");
-        Hyperlink delete = new Hyperlink("Delete plan");     
+        Hyperlink delete = new Hyperlink("Delete plan");   
+        Hyperlink resetCourses = new Hyperlink("Reset courses");
         
         VBox suggestedCourses = new VBox();
         Text suggestTitle = new Text("");
@@ -237,12 +245,14 @@ public class PlanAppUi extends Application {
         courseCollection.setPadding(new Insets(10, 10, 10, 10));
         
         List<CheckBox> boxes = new ArrayList<>();
- 
+                
+        // Save button function
         saveButton.setOnAction((ActionEvent push) -> {
+            List<Course> sysCourses = planService.allCourses();
+            List<Course> nextCourses = new ArrayList<>();
             suggestedCourses.getChildren().clear();
             suggestedCourses.getChildren().add(suggestTitle);
             suggestTitle.setText("Suggested courses: ");
-            nextCourses.clear();
             
             for (CheckBox c : boxes) {
                 String pcs[] = c.getText().split(": ");
@@ -250,16 +260,12 @@ public class PlanAppUi extends Application {
                 
                 if (c.isSelected() && !planService.selectedCourses().contains(cc)) {
                     planService.addCourse(cc);
-                } else if (c.isSelected() && planService.selectedCourses().contains(cc)) {
-                    continue;
-                } else {
-                    if (planService.selectedCourses().contains(cc)) {
-                        planService.removeCourse(cc);
-                    }
-                }
+                } else if (!c.isSelected() && planService.selectedCourses().contains(cc)) {
+                    planService.removeCourse(cc);
+                } 
             }            
             
-            for (Course c : courses) {
+            for (Course c : sysCourses) {
                 List<Course> prerequisites = c.getPrerequisites();
                 List<Course> completed = planService.selectedCourses();
                                 
@@ -285,18 +291,27 @@ public class PlanAppUi extends Application {
             }
         });
         
+        
+        // Sidemenu settings
         menuTitle.setFont(Font.font("Arial", FontWeight.BOLD, 14));
         VBox.setMargin(menuTitle, new Insets(0, 0, 0, 2));
         studySubjects.getChildren().add(menuTitle);
         
         Hyperlink csLink = new Hyperlink("TKT-Courses");
-        Hyperlink prerequisitesTKT = new Hyperlink("TKT-Prerequisites");
+        Hyperlink modCsLink = new Hyperlink("Modify TKT-Courses");
         
         VBox.setMargin(csLink, new Insets(0, 0, 0, 8));
-        VBox.setMargin(prerequisitesTKT, new Insets(0, 0, 0, 8));
+        VBox.setMargin(modCsLink, new Insets(0, 0, 0, 8));
         
+        // Tools for planning
         csLink.setOnAction((ActionEvent push) -> {
+            List<Course> sysCourses = planService.allCourses();
+            boxes.clear();
+            buttonBar.getChildren().clear();
+            buttonBar.add(saveButton, 0, 0);
             courseCollection.getChildren().clear();
+            suggestedCourses.getChildren().clear();
+            
             if (planService.selectedCourses().isEmpty()) {
                 planInstruction.setText("Start by selecting all completed courses");
             } else {
@@ -305,7 +320,7 @@ public class PlanAppUi extends Application {
             
             courseCollection.getChildren().add(planInstruction);
             
-            for (Course c : courses) {
+            for (Course c : sysCourses) {
                 CheckBox box = new CheckBox(c.toString());
             
                 if (planService.selectedCourses().contains(c)) {
@@ -316,22 +331,165 @@ public class PlanAppUi extends Application {
                 courseCollection.getChildren().add(box);
             }
             
-            courseCollection.getChildren().add(saveButton);
+            courseCollection.getChildren().add(buttonBar);
+        });
+        
+        Button deleteCourse = new Button("Delete");
+        Button createCourse = new Button("Create new");
+        
+        // Course deletion from system
+        deleteCourse.setOnAction((ActionEvent push) -> {            
+            for (CheckBox cb : boxes) {
+                String[] pcs = cb.getText().split(": ");
+                Course c = planService.findCourse(pcs[0]);
+                
+                if (cb.isSelected()) {
+                    planService.sysDeleteCourse(c);
+                }
+            }
+            
+            modCsLink.fire();
+        });
+        
+        // Course creation
+        Label newCode = new Label("Code: ");
+        Label newName = new Label("Name: ");
+        
+        TextField newCourseCode = new TextField();
+        TextField newCourseName = new TextField();
+        
+        Button createNewCourse = new Button("Create");
+        Button cancelCreatingCourse = new Button("Cancel");
+        
+        GridPane courseCreatePane = new GridPane();
+        courseCreatePane.setPadding(new Insets(10, 10, 10, 10));
+        courseCreatePane.add(newCode, 0, 0);
+        courseCreatePane.add(newCourseCode, 1, 0);
+        courseCreatePane.add(newName, 0, 1);
+        courseCreatePane.add(newCourseName, 1, 1);
+        
+        // START creating a new course
+        createCourse.setOnAction((ActionEvent push) -> {
+            planNotification.setText("");
+            planNotification1.setText("");
+            newCourseCode.setText("");
+            newCourseName.setText("");
+            
+            buttonBar.getChildren().clear();
+            courseCollection.getChildren().clear();
+            planInstruction.setText("New course:");
+            
+            courseCollection.getChildren().add(planInstruction);
+            courseCollection.getChildren().add(courseCreatePane);            
+            
+            // Existing courses as a list for choosing prerequisites
+            suggestTitle.setText("Choose prerequisites for the course:");
+            suggestedCourses.getChildren().add(suggestTitle);
+            for (CheckBox cb : boxes) {
+                suggestedCourses.getChildren().add(cb);
+            }
+            
+            buttonBar.add(createNewCourse, 0, 0);
+            buttonBar.add(cancelCreatingCourse, 1, 0);
+            
+            courseCollection.getChildren().add(buttonBar);
+            courseCollection.getChildren().add(planNotification);
+            courseCollection.getChildren().add(planNotification1);
+        });
+        
+        // FINISH creating a new course
+        createNewCourse.setOnAction((ActionEvent push) -> {
+            planNotification.setText("");
+            planNotification1.setText("");
+            
+            boolean codeLengthErr = false;
+            boolean nameLengthErr = false;
+            List<Course> prerequisites = new ArrayList<>();
+            
+            for (CheckBox cb : boxes) {
+                String[] pcs = cb.getText().split(": ");
+                Course c = planService.findCourse(pcs[0]);
+                
+                if (cb.isSelected()) {
+                    prerequisites.add(c);
+                }
+            }
+            
+            if (newCourseCode.getText().length() > 13 || newCourseCode.getText().length() < 5) {
+                planNotification.setFill(Color.FIREBRICK);
+                planNotification.setText("Code length must be between 5-13");
+                codeLengthErr = true;
+            } 
+            
+            if (newCourseName.getText().length() > 50 || newCourseName.getText().length() < 3) {
+                planNotification1.setFill(Color.FIREBRICK);
+                planNotification1.setText("Name length must be between 3-50");
+                nameLengthErr = true;
+            }
+            
+            if (!codeLengthErr && !nameLengthErr) {
+                Course nc = new Course(newCourseCode.getText(), newCourseName.getText());
+                if (!planService.allCourses().contains(nc)) {
+                    if (planService.sysCreateCourse(nc, prerequisites)) {
+                        planNotification.setFill(Color.GREEN);
+                        planNotification.setText("Course created successfully");
+                        modCsLink.fire();
+                    } else {
+                        planNotification.setFill(Color.FIREBRICK);
+                        planNotification.setText("Something went wrong");
+                    }
+                } else {
+                    planNotification.setFill(Color.FIREBRICK);
+                    planNotification.setText("Course already exists");
+                }
+            }
+        });
+        
+        // Cancel-function inside course creation
+        cancelCreatingCourse.setOnAction((ActionEvent push) -> {
+            newCourseCode.clear();
+            newCourseName.clear();
+            planNotification.setText("");
+            planNotification1.setText("");
+            
+            modCsLink.fire();
+        });
+                
+        
+        // Link for TKT-course modifications      
+        modCsLink.setOnAction((ActionEvent push) -> {
+            List<Course> sysCourses = planService.allCourses();
+            boxes.clear();
+            buttonBar.getChildren().clear();
+            courseCollection.getChildren().clear();
+            suggestedCourses.getChildren().clear();
+            planInstruction.setText("Delete and Create Courses");
+            courseCollection.getChildren().add(planInstruction);
+            
+            for (Course c : sysCourses) {
+                CheckBox box = new CheckBox(c.toString());
+                boxes.add(box);
+                courseCollection.getChildren().add(box);
+            }
+            
+            if (!boxes.isEmpty()) {
+                buttonBar.add(deleteCourse, 1, 0);
+            }
+            buttonBar.add(createCourse, 0, 0);
+            courseCollection.getChildren().add(buttonBar);       
             courseCollection.getChildren().add(planNotification);
         });
         
-        prerequisitesTKT.setOnAction((ActionEvent push) -> {
-            courseCollection.getChildren().clear();
-            suggestedCourses.getChildren().clear();
-            planInstruction.setText("Prerequisites graph for TKT-courses: (Coming soon)");
-            courseCollection.getChildren().add(planInstruction);
-        });
-        
         studySubjects.getChildren().add(csLink);
-        studySubjects.getChildren().add(prerequisitesTKT);
+        studySubjects.getChildren().add(modCsLink);
         studySubjects.setStyle("-fx-background-color: #CACACA");
-                
+        
+        // Logout function
         logout.setOnAction((ActionEvent push) -> {
+            newCourseCode.clear();
+            newCourseName.clear();
+            planNotification.setText("");
+            planNotification1.setText("");
             suggestedCourses.getChildren().clear();
             courseCollection.getChildren().clear();
             
@@ -342,22 +500,56 @@ public class PlanAppUi extends Application {
             stg.setScene(login);
         });
         
+        // Plan deletion
+        Alert delConf = new Alert(AlertType.CONFIRMATION);
+        delConf.setTitle("Plan Deletion");
+        delConf.setHeaderText("Confirmation Required");
+        delConf.setContentText("This will permanently delete your plan");
+                        
         delete.setOnAction((ActionEvent push) -> {
+            newCourseCode.clear();
+            newCourseName.clear();
+            planNotification.setText("");
+            planNotification1.setText("");
             logUsername.setText("");
-            if (planService.deletePlan()) {
-                logNotification.setFill(Color.GREEN);
-                logNotification.setText("Plan deleted successfully");
-            } else {
-                logNotification.setFill(Color.FIREBRICK);
-                logNotification.setText("Something went wrong");
-            }
             
-            stg.setX((screenBounds.getWidth() - 500) / 2);
-            stg.setY((screenBounds.getHeight() - 200) / 2);
-            stg.setScene(login);
+            Optional<ButtonType> confResult = delConf.showAndWait();
+            
+            if (confResult.get() == ButtonType.OK) {
+                if (planService.deletePlan()) {
+                    logNotification.setFill(Color.GREEN);
+                    logNotification.setText("Plan deleted successfully");
+                } else {
+                    logNotification.setFill(Color.FIREBRICK);
+                    logNotification.setText("Something went wrong");
+                }
+                
+                stg.setX((screenBounds.getWidth() - 500) / 2);
+                stg.setY((screenBounds.getHeight() - 200) / 2);
+                stg.setScene(login);
+            }             
         });
         
-        links.getChildren().addAll(logout, delete);
+        // System course reset
+        resetCourses.setOnAction((ActionEvent push) -> {
+            newCourseCode.clear();
+            newCourseName.clear();
+            planInstruction.setText("");
+            planNotification.setText("");
+            planNotification1.setText("");
+            logout.fire();
+            planService.sysCourseReset();
+            try {
+                init();
+                logNotification.setFill(Color.GREEN);
+                logNotification.setText("Courses have been reset");
+            } catch (Exception e) {
+                logNotification.setFill(Color.FIREBRICK);
+                logNotification.setText("Something went wrong during course reset");
+            }
+        });
+        
+        links.getChildren().addAll(logout, delete, resetCourses);
         actions.getChildren().addAll(links);
         AnchorPane.setRightAnchor(links, 10.0);
         
@@ -368,19 +560,21 @@ public class PlanAppUi extends Application {
         topBar.setRight(actions);
         topBar.setLeft(welcome);        
         
+        planFlex.setContent(courseCollection);
+        suggFlex.setContent(suggestedCourses);
+        
         planGrid.setLeft(studySubjects);
-        planGrid.setRight(suggestedCourses);
-        planGrid.setCenter(courseCollection);
+        planGrid.setRight(suggFlex);
         planGrid.setTop(topBar);
                 
         plan = new Scene(planGrid, 1000, 600);
         
         
         // Run application:
-        stg.setX((screenBounds.getWidth() - 500) / 2);
-        stg.setY((screenBounds.getHeight() - 200) / 2);
         stg.setTitle("PlanApp");
         stg.setScene(login);
+        stg.setX((screenBounds.getWidth() - 500) / 2);
+        stg.setY((screenBounds.getHeight() - 200) / 2);
         stg.show();       
     }
